Separate React and non-React components of date/time package
