+++ packages/datetime/README.md
@@ -17,11 +17,10 @@ npm i react-querybuilder @react-querybuilder/datetime
 
 ## Usage
 
-<!--
 To enable the date and time functionality of a query builder, nest the `QueryBuilder` element under `QueryBuilderDateTime`.
 
 ```tsx
-import { QueryBuilderDateTime } from '@react-querybuilder/datetime';
+import { QueryBuilderDateTime } from '@react-querybuilder/datetime/ui';
 import { useState } from 'react';
 import { QueryBuilder, RuleGroupType } from 'react-querybuilder';
 
@@ -40,7 +39,14 @@ const App = () => {
   );
 };
 ```
--->
+
+All React exports live in the `/ui` entry point (`@react-querybuilder/datetime/ui`); the root entry is framework-agnostic, exporting only rule processors, utilities, and types.
+
+- **`QueryBuilderDateTime`** — the context provider shown above. Accepts optional config props (`anchors`, `units`, `toggleLabels`, `modeController`, `dateTimeAPI`) to customize the relative date/time editor.
+- **`RelativeDateTimeValueEditor`** — the value editor itself, for use directly via `controlElements` when you'd rather not wrap with the provider.
+- **Mode controllers** — `toggleModeController` (the default absolute/relative toggle button) and `createOperatorModeController` (derives the mode from the rule's operator), plus the `withRelativeOperators` field helper.
+
+See the [Components documentation](https://react-querybuilder.js.org/docs/datetime#components) for relative values, mode controllers, and configuration.
 
 To add date/time functionality to `formatQuery`, import a rule processor from `@react-querybuilder/datetime/dayjs`, `@react-querybuilder/datetime/date-fns`, or `@react-querybuilder/datetime/luxon`—depending on which library you wish to use or are already using—and pass it to the `ruleProcessor` option.
