+++ packages/datetime/dev/main.tsx
@@ -2,7 +2,7 @@ import { DevLayout, useDevApp } from '@rqb-devapp';
 import * as React from 'react';
 import { createRoot } from 'react-dom/client';
 import { QueryBuilder } from 'react-querybuilder';
-import { QueryBuilderDateTime } from '../src/QueryBuilderDateTime';
+import { QueryBuilderDateTime } from '../src/ui/QueryBuilderDateTime';
 import './styles.scss';
 
 const App = () => {
