+++ packages/datetime/src/getDatetimeRuleProcessorCypher.ts
@@ -1,5 +1,5 @@
-import type { RuleProcessor } from 'react-querybuilder';
-import { defaultRuleProcessorCypher, lc, toArray } from 'react-querybuilder';
+import type { RuleProcessor } from '@react-querybuilder/core';
+import { defaultRuleProcessorCypher, lc, toArray } from '@react-querybuilder/core';
 import type { RQBDateTimeLibraryAPI } from './types';
 import {
   isISOStringDateOnly,
