+++ packages/datetime/src/index.date-fns.ts
@@ -4,7 +4,7 @@
  * @module date-fns
  */
 
-import type { RuleProcessor, ValueProcessorByRule } from 'react-querybuilder';
+import type { RuleProcessor, ValueProcessorByRule } from '@react-querybuilder/core';
 import { getDatetimeJsonLogicOperations } from './getDatetimeJsonLogicOperations';
 import { getDatetimeRuleProcessorCEL } from './getDatetimeRuleProcessorCEL';
 import { getDatetimeRuleProcessorCypher } from './getDatetimeRuleProcessorCypher';
