+++ packages/datetime/src/index.dayjs.ts
@@ -4,7 +4,7 @@
  * @module dayjs
  */
 
-import type { RuleProcessor, ValueProcessorByRule } from 'react-querybuilder';
+import type { RuleProcessor, ValueProcessorByRule } from '@react-querybuilder/core';
 import { getDatetimeJsonLogicOperations } from './getDatetimeJsonLogicOperations';
 import { getDatetimeRuleProcessorCEL } from './getDatetimeRuleProcessorCEL';
 import { getDatetimeRuleProcessorCypher } from './getDatetimeRuleProcessorCypher';
