+++ packages/datetime/src/index.jsdate.ts
@@ -4,7 +4,7 @@
  * @module jsdate
  */
 
-import type { RuleProcessor, ValueProcessorByRule } from 'react-querybuilder';
+import type { RuleProcessor, ValueProcessorByRule } from '@react-querybuilder/core';
 import { getDatetimeJsonLogicOperations } from './getDatetimeJsonLogicOperations';
 import { getDatetimeRuleProcessorCEL } from './getDatetimeRuleProcessorCEL';
 import { getDatetimeRuleProcessorCypher } from './getDatetimeRuleProcessorCypher';
